// src/components/imports/models/transaction.model.ts
export interface Transation {
    StoreCode: string;
    Loc1: number;
    ShrmDesc: string;
    InvoiceDate: string;
    SaleType: string;
    InvoiceNumber: number;
    RefInvoiceNumber: number;
    MobileNo: string;
    Product: string;
    Mrp: number;
    SalesQty: number;
    SalesValue: number;
    Discount: number;
    SalesMgr: string;
    State: string;
    RegMgr: string;
}